GetMappedFieldName()
Given a Type and a field name, this method will return the corresponding
field name on the datastore. It is assumed that a DataContract has been
used to decorate the Type. If it hasn't, than the same field name is
returned (assuming the Type field and the datastore field names match).